Rename Tree.Error to ErrorRegion
